refac: 関数の切り出し、型定義の移動
