+++ app/view/[teamName]/[recordId]/page.tsx
@@ -2,50 +2,7 @@ import { getRecordDetailsByRecordId } from '@/lib/api/record'
 import ScoreBoard from './components/scoreBoard'
 import StoneCoordinates from './components/stoneCoordinates'
 import { DeleteRecordButton } from './components/DeleteRecordButton'
-import { RecordDetail } from '@/lib/api/types/model'
-
-export function extractTeamsScoreData(record: RecordDetail): TeamScore {
-  const myTeamName = 'My Team' // 自チーム名を適切に設定してください
-  const opponentTeamName = record.enemy_team_name
-
-  const myTeamScores: (number | string)[] = ['']
-  const opponentTeamScores: (number | string)[] = ['']
-  let myTeamTotal = 0
-  let opponentTeamTotal = 0
-
-  record.ends_data.forEach((end) => {
-    if (end.score >= 0) {
-      myTeamScores.push(end.score)
-      opponentTeamScores.push(0)
-      myTeamTotal += end.score
-    } else {
-      myTeamScores.push(0)
-      opponentTeamScores.push(Math.abs(end.score))
-      opponentTeamTotal += Math.abs(end.score)
-    }
-  })
-
-  // スコアが10エンドに満たない場合、0で埋める
-  while (myTeamScores.length < 11) {
-    myTeamScores.push(0)
-    opponentTeamScores.push(0)
-  }
-
-  return {
-    friend: {
-      team: myTeamName,
-      color: 'red',
-      scores: myTeamScores,
-      total: myTeamTotal
-    },
-    enemy: {
-      team: opponentTeamName,
-      color: 'yellow',
-      scores: opponentTeamScores,
-      total: opponentTeamTotal
-    }
-  }
-}
+import { extractTeamsScoreData } from '@/utils/func/score'
 
 export default async function RecordPage({ params }: { params: { teamName: string; recordId: string } }) {
   console.log(params)
