+++ lib/api/record.ts
@@ -1,6 +1,6 @@
+import { getRecordDetailsByRecordIdResponse, getRecordIndicesByTeamIdResponse } from '@/app/@types/response'
 import { apiClient } from '@/utils/api/api'
 import { cache } from 'react'
-import { getRecordDetailsByRecordIdResponse, getRecordIndicesByTeamIdResponse } from './types/response'
 
 export const getRecordsByTeamId = cache(async (teamId: string): Promise<getRecordIndicesByTeamIdResponse> => {
   const res = await apiClient.get<getRecordIndicesByTeamIdResponse>(`/auth/records/${teamId}`)
